requirements.txt
flask==3.0.3
pypdf==4.3.1
python-docx==1.1.2
